confirm with the user to create a file when it doesn't exist
Previously a file would just be created, but with this patch the user
will be asked for confirmation when trying to open a file that doesn't
exist yet. If they don't answer yes, no file shall be created.

+++ edit.py
@@ -1,4 +1,4 @@
-import sys, signal
+import sys, signal, os
 
 print "=================="
 print "welcome to [chat-edit]"
@@ -37,6 +37,10 @@ class ChatEdit:
         args = " ".join(words[1:])
         
         if command in openwords:
+            if not os.path.isfile(args):
+                if not confirm("create " + args + "?"):
+                    msg("filesystem not touched")
+                    return
             self.file = open(args, "w+")
             self.contents = self.file.read()
             self.changes = False
